ir_prox: Improve filtering & reduce bandwidth
Check that the received pulse is within a symbol range.
Increase gain and offset to lower the error rate when the robots are
moving.

The "noise" is about +/- 2 clock cycle when still (which would lead to
13 bits of useable data in 4ms).
But as soon as the robot is moving (ie, the intensity change over time),
the noise is much higher.
This is an electronic problem and cannot be worked around in software
(checked with a scope ....)