motor: add calibration support
Some motor have different specifications (magnets ?).
Since we control the motor's speed based on the back EMF, we need to be
able to change the applied back EMF based on a calibration specific to
each robot/motors.

This commit adds two settings (8,9) in order to have a controllable gain
on the correction:

PID target = aseba target * setting / 256.
aseba speed = physical back EMF * 256 / setting.

With this commit we can automate the motor calibration with an external
motor forcing the robot's motor's speed at an idealized speed of 256 and
reading the actual back EMF from the motor. This value can directly be
written in the settings to compensate the back EMF difference between
motors.